requirements.txt
phidata
yfinance
googlesearch-python
pycountry
packaging
fastapi
uvicorn
groq
python-dotenv
Flask